// src/app/interfaces/lecture-publisher/workFlowProd.ts
export interface sopProdDTO {
  fSOPID: number;
  fPubId: number;
  fSopName: string;
  fPubName: string;
  fPubContent: string;
  fPubSopImagePath: string;
  fPrice: number;
  fSalePoints: number;
  fReleaseStatus: string;
  fJobItemId: number;
  fIndustryId: number;
  fCompanySize: string;
  fReleaseTime: string;
  fSopDescription: string;
  quantity: number;
  fPubSopImageUrl: string;

}